Fix types

+++ tensorly/tenalg/tests/test_n_mode_product.py
@@ -55,8 +55,8 @@ def test_mode_dot():
     res = mode_dot(X, U, 1)
     assert_array_equal(true_res, res)
     # Test with a third order tensor
-    X = T.tensor(np.arange(24).reshape((3, 4, 2)))
-    v = T.tensor(np.arange(4))
+    X = T.tensor(np.arange(24, dtype=float).reshape((3, 4, 2)))
+    v = T.tensor(np.arange(4, dtype=float))
     true_res = ([[ 28,  34],
                  [ 76,  82],
                  [124, 130]])
@@ -110,7 +110,7 @@ def test_multi_mode_dot():
     res = multi_mode_dot(X, U, [0, 1])
     assert_array_equal(true_res, res)
 
-    X = T.tensor(np.arange(12).reshape((3, 4)))
+    X = T.tensor(np.arange(12, dtype=float).reshape((3, 4)))
     U = T.tensor(np.random.random((3, 5)))
     res_1 = multi_mode_dot(X, [U], modes=[0], transpose=True)
     res_2 = T.dot(T.transpose(U), X)
